Schema walking and manipulation utilities.
This package currently provides several functions which will
be split into at least two packages in subsequent commits.

It provides several utility functions relating to walking
over a schema and/or its subschemas and applying callbacks
before and/or after each step of the walk.

It also provides utlities to find specific link description objects
within a schema, as well as work with several Cloudflare-specfic
extensions to JSON Schema / Hyper-Schema.