Fix wrong variable and typo
We wanted path.length, not length.  The unit tests failed to
catch this because the test environment somehow defines
length globally as 0 (this is not the case when called from
elsewhere), and path is always 2 if we exectute this branch
at all.  And slicing 1 or -1 has the same effect.

This is such a weird confluence of errors that, given that we
have no paths that would be longer than 2 in any draft of
JSON Schema, trying to get this tested would require way more
effort than seems worthwhile.  We would have to instrument the
walker or re-work the code to allow testing this section in
isolation.