Add functions for collapsing allOf'd schemas
Collapsing schemas from an "allOf" is not a simple merge.  Keywords
have to be combined based on their own semantics, and some cannot
be collapsed.

This adds the simplest functions and sets up the vocabularies.
Commits will follow immediately adding more complex functions
and pulling them together into a callback for use with the
json-schema-walker package.