Make things actually work.
The module wasn't actually imported to index.js, and the "this"
trick for test mocking didn't work when the package is used
normally (which, in retrospect, should have been obvious).

For lack of a better idea, use the function that we want to mock
from "this" if it's available, otherwise use the normal one.
I can't say I like this idea but it unblocks things for now.

Also, fix broken string template.