Basic example roll-up code
This is a callback for use with the json-schema-walker pacakge.
It rolls up example values with the expectation that "allOf"s have
all been collapsed (although we can add support for rolling up
through "allOfs" in the future if desired).

Since JSON Schema allows for ambiguous and/or unspecified
types, there are some heuristics for detecting the likely type
and deciding how to roll up examples.  We might want to tweak
these but they're a good start.

Unlike the rollup code in json-schema-example-loader, which this
replaces, the examples remain regualar data structures and not
JSON Strings, which then get turned into escaped JSON strings when
the data is stringified later.  Some part of the system closer
to the display, quite likely the react components, will decide how
to format the examples rather than have it be done during loading.

There are no doubt some other minor differences either due to this
code or the allOf-collapsing code (which takes the first example
it finds, while json-schema-example-loader would take the last).
It also may be true that this code keeps an upper level example
in some situations where json-schema-example-loader would
overwrite it with a lower-level example.

I think that the behavior that objects and arrays never look
at "default" for an example is the same, but I'm not 100%
certain it's the same along all old code paths, nor am I 100%
certain that it's the right behvior.  But it's a starting point.

The goal here is to define reasonable and consistent behavior
rather than mimic the current set-up, which is slightly different
for root schemas vs subschemas anyway so we can't perfectly mimic it.

For feature parity, we still need to support "minItems", which will
be done as a separate commit.