Integration rather than unit tests.
The amount of effort and confusing techniques involved in mocking
was not even close to worthwhile for these tests.  If the code
being mocked was more complex it would be worthwhile to look for
a helper library for mocking, but this is simple, same-module,
already-tested code so just test the whole integrated module.
It's vastly easier to understand.

Also added some test cases that did not show up through the
coverage report, and found a bug.  Although it's debatable
whether the intended behavior is really desirable so I also
added a TODO to reconsider it.  We'll need to understand use
cases a bit better first, I think.