Add support for minimums and maximums
Includes support for the excessively complicated draft-04
boolean exclusivity modifiers.  These were replaced in draft-06
with a simpler system.

Since we only handle keywords specially when they are in both the
parent schema and subschema, there is a weird special case of
an "exclusive" boolean being present only in the parent schema,
without the keyword that it modifies.  Add some special code to
just remove that, as it has no effect anyway.  It cannot affect
the value from the subschema as exclusivity flags *MUST* only
modify immediately adjacent keywords.  (see? complicated)