Switch to leaving out undefined examples
While writing the system test for example rollup, it became
apparent that putting JavaScript "undefined" literals as
example placeholder creates problems.  They translate to JSON
"null" literals, but an example of null is not the same as a
missing example.

Rework the logic to avoid adding an example field when no
example rollup can be constructed.  Do not fill in array positions
or required object fields with undefined; instead do not roll up
those arrays or objects.

Now that doca displays nested structures, the individual examples
can be seen when the nested documentation is displayed.