System tests for example rollup
This uncovered a missing case in the example rollup where a usable
example from a *Of keyword was ignored if type was set explicitly
without a local example (adjacent to the *Of).

Some of the array min/max system test data was changed to be a
better case for the example rollup, in particular not blocking
the complex minItems case by providing an example above the
"items" keyword.