Add curl example system test
While doing so, noticed that there is no space before the line
continuation backslash, which is a change from the current display.
And not a good one, so this puts the space back in.

This does not cover the cfRequestHeaders case as the walker code
does not understand that keyword yet.