Remove non-standard URI Template support
We have previously used JSON Pointers as URI Template
variables for the purpose of finding example values.

This had a number of problems:

* "#" is a reserved character in RFC 6570 URI Templates
* There was a wierd id vs identifier hack
* There's no relation between the variable and the
  actual instance data

This change drops support for this approach entirely.

For now, we are keeping the dotted notation feature,
which is left open to applications by RFC 6570.  We
may replace this with a backport of "templatePointers"
from draft-07 in the future.

There are several outstanding concerns:

* Not all URI Templates in our schemas have an instance
  field for every template variable
* Sometimes it may be preferable to display a different
  variable name than the instance variable

Both of these can be addressed by backporting limited
support for "hrefSchema" from later drafts of Hyper-Schema,
which will be done in a future commit.

Note that the DELETE endpoint was added to the system test
fixtures to cover the JSON Pointer template variable case.
After deciding to drop that feature, the added endpoint
still adds system test coverage for cfRequestHeaders so I
left it in.