Explain cfPrivate and cfOmitFromExample clearly
This change only effects comments.

cfOmitFromExample should be explained in the docs for the rollup
function, even thought it does exactly what it says.  It's worth
noting that examples are still sometimes useful for those fields
even when not rolled up.

cfPrivate does not affect example creation, but that is not
necessarily obvious so make a note of it.