Instance and input data for template resolution
Rather than just passing one data source, allow for two, and
indicate which is instance vs input so that we can handle the
two separately as we add support for draft-07 features.

For now, just resolve from them in the correct order.  Defer
validation of input against hrefSchema and more complex
handling that involves other draft-07 keywords to the future.