Initial working default theme files
This is basically our internal theme with almost all components
removed.  It just shows the placeholder introduction (with
base URI from the config) and the raw JSON Schema data.
This is good enough for debugging, which is good enough for now.

The App and Head components and at least one CSS file constitute
the interface with the scaffolded Doca app.