Use walker vocabularies in transform
These were not upated when the walker vocabulary concept was
added.

While we're at it, make a few changes to make sure that the behavior
of the vocabularies is pretty similar (e.g. CLOUDFLARE_DOCA includes
the standard keywords for both), and fill in missing drafts.

The whole vocabulary thing is a bit awkward, but is being dramatically
refined and reworked in draft-08 so we'll tackle a better design
once the draft-08 proposal is finalized.